完成模型加载 GUI 界面，热加载等全部功能 - 完成模型描述文件加载功能 - 完成模型切换 GUI 界面 - 模型切换 GUI 界面文本描述，翻页功能全部完成 - 模型热加载重载功能完成 - 修改部分其他内容

+++ src/main/java/com/github/tartaricacid/touhoulittlemaid/client/model/pojo/CubesItem.java
@@ -12,7 +12,7 @@ public class CubesItem {
     private boolean mirror;
 
     @SerializedName("size")
-    private List<Integer> size;
+    private List<Float> size;
 
     @SerializedName("origin")
     private List<Float> origin;
@@ -25,7 +25,10 @@ public class CubesItem {
         return mirror;
     }
 
-    public List<Integer> getSize() {
+    /**
+     * 基岩版这货居然可以为浮点数，服了
+     */
+    public List<Float> getSize() {
         return size;
     }
